requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastobo==0.12.2
fhir.resources==7.0.1
fhirpathpy==0.1.2
frozenlist==1.3.3
fsspec==2023.6.0
idna==3.4
multidict==6.0.4
networkx==3.1
numpy==1.25.0
nxontology==0.5.0
pandas==2.0.2
pronto==2.5.4
pydantic==1.10.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
